chore: 데이터베이스 모델 분리
