+++ server/tsconfig.path.json
@@ -2,7 +2,8 @@
   "compilerOptions": {
     "baseUrl": ".",
     "paths": {
-      "@config": ["config"]
+      "@config": ["config"],
+      "@db": ["db"]
     }
   }
 }
